[CHORE] dto 분리, getAllComment 분리
