--- src/interface/comment/CommentResponseDto.ts
@@ -1,8 +1,0 @@
-export interface CommentResponseDto {
-  nickName: string;
-  photo: string | null;
-  content: string | null;
-  emoji: number | null;
-  createdAt: Date;
-  updatedAt: Date;
-}
